backend/requirements.txt
blinker==1.8.2
boto3==1.34.147
botocore==1.34.147
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fsspec==2024.6.1
gunicorn==22.0.0
huggingface-hub==0.24.2
humanfriendly==10.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.18.1
packaging==24.1
protobuf==5.27.2
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
s3transfer==0.10.2
safetensors==0.4.3
six==1.16.0
sympy==1.13.1
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.43.2
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
